Fixed Bugs, further refactoring
