+++ scripts/bfasst/compare_waveforms/File_Parsing/parse_diff.py
