+++ scripts/bfasst/compare_waveforms/Templates/structs.py
