+++ scripts/bfasst/compare_waveforms/Templates/template.tcl
